+++ src/Components/Carousel/NavArrow/NavArrow.styled.tsx
@@ -1,0 +1,20 @@
+import { SvgIcon } from 'Components/SvgIcon'
+import styled from 'styled-components'
+
+export const StyledNavArrowButton = styled.button`
+  background-color: transparent;
+  border: 0;
+  margin: 0;
+  outline: 0;
+  padding: 0.75rem 1rem;
+`
+
+export const StyledArrowIcon = styled(SvgIcon)<{
+  disabled?: boolean
+}>`
+  svg * {
+    fill: ${({ theme: { colors }, disabled }) =>
+      // disabled ? 'transparent' : colors.GREY_250};
+      disabled ? 'transparent' : '#BBC0C3'};
+  }
+`