+++ src/Components/CarouselWithReactCarousel/NavArrow/NavArrow.styled.tsx
@@ -1,0 +1,21 @@
+import styled from 'styled-components'
+
+import { ButtonBasic } from '../../Button'
+import { SvgIcon } from '../../SvgIcon'
+
+export const StyledNavArrowButton = styled(ButtonBasic)`
+  background-color: transparent;
+  border: 0;
+  margin: 0;
+  outline: 0;
+  padding: 0.75rem 1rem;
+`
+
+export const StyledArrowIcon = styled(SvgIcon)<{
+  disabled?: boolean
+}>`
+  svg * {
+    fill: ${({ theme: { colors }, disabled }) =>
+      disabled ? 'transparent' : colors.GREY_250};
+  }
+`