+++ src/utils/ColorControl.ts
@@ -1,0 +1,9 @@
+import { COLOR } from 'styles/colors'
+
+export const ColorControl = {
+  options: Object.keys(COLOR),
+  mapping: COLOR,
+  control: {
+    type: 'select',
+  },
+}